Add image

+++ model/actor/actor.php
@@ -1,6 +1,7 @@
 <?php
 
 require_once $_SERVER['DOCUMENT_ROOT'] . '/media/global/utils.php';
+require_once $_SERVER['DOCUMENT_ROOT'] . '/media/model/content/content.php';
 
 class actor{
 	
@@ -46,7 +47,7 @@ class actor{
 
 	static function getList($bOption=null,$bNone=null,$bAll=null,$options=null){
 		global $bdd;
-        $sql="select idActor,firstName,lastName from actor";
+        $sql="select idActor,firstName,lastName,image from actor";
         $whereoption=" where idActor>0";
         $search="";
         $orderby=" order by firstName";
@@ -94,7 +95,7 @@ class actor{
 
 	function getActor($id){
 		global $bdd;
-		$req = $bdd->prepare('select idActor,firstName,lastName from actor where idActor=:id');
+		$req = $bdd->prepare('select idActor,firstName,lastName,image from actor where idActor=:id');
 		$req->bindParam(":id",$id);
 		$result=$req->execute();
 		if ($result==1){
@@ -102,6 +103,7 @@ class actor{
 			$this->idActor=$data['idActor'];
 			$this->firstName=$data['firstName'];
 			$this->lastName=$data['lastName'];
+			$this->image=$data['image'];
 		}else{
 			$data=init();
 		}
@@ -109,21 +111,49 @@ class actor{
 		return $data;
 	}
 
+	static function getImage($id){
+		global $bdd;
+		$req = $bdd->prepare('select image from actor where idActor=:id');
+		$req->bindParam(":id",$id);
+		$result=$req->execute();
+		if ($result==1){
+			$data = $req->fetch();
+			return $data['image'];
+		}else{
+			return '';
+		}
+	}
+
 	static function delete($id){
 		global $bdd;
-		$req = $bdd->prepare('delete from actor where idActor=:id');
-		$req->bindParam(":id", $id, PDO::PARAM_INT);
-		$result=$req->execute();
+		$image=actor::getImage($id);
+		if (strlen($image)>0) {
+			$result=unlink($_SERVER['DOCUMENT_ROOT'] . $image);
+			if ($result==true) {
+				$result = unlink($_SERVER['DOCUMENT_ROOT'] . 'small_' . $image);
+			}
+		}
+		if ($result==true) {
+			$req = $bdd->prepare('delete from actor where idActor=:id');
+			$req->bindParam(":id", $id, PDO::PARAM_INT);
+			$result = $req->execute();
+			if ($result == 1) {
+				print 'content delete';
+				$result = content::deleteContentActor($id);
+			}
+		}
 		return (int)$result;
 	}
 
 	function insert(){
 		global $bdd;
-        $search=setSearchString($this->firstName.$this->lastName);
-		$req = $bdd->prepare('insert into actor (firstName,lastName,search) values (:firstname,:lastname,:search)');
+		$image="";
+		$search=setSearchString($this->firstName.$this->lastName);
+		$req = $bdd->prepare('insert into actor (firstName,lastName,search,image) values (:firstname,:lastname,:search,:image)');
 		$req->bindParam(":firstname", $this->firstName, PDO::PARAM_STR);
 		$req->bindParam(":lastname", $this->lastName, PDO::PARAM_STR);
         $req->bindParam(":search", $search, PDO::PARAM_STR);
+		$req->bindParam(":image", $image, PDO::PARAM_STR);
 		$result=$req->execute();
 		$this->idActor = $bdd->lastInsertId();
 		return (int)$result;
@@ -152,6 +182,15 @@ class actor{
 		return (int)$result;
 	}
 
+	static function setImage($id,$image){
+		global $bdd;
+		$req = $bdd->prepare('update actor set image=:image where idActor=:id');
+		$req->bindParam(":id", $id, PDO::PARAM_INT);
+		$req->bindParam(":image", $image, PDO::PARAM_STR);
+		$result=$req->execute();
+		return (int)$result;
+	}
+
 	static function import(){
 		$row = 1;
 		if (($handle = fopen("C:/wamp/www/media/actor.csv", "r")) !== FALSE) {
@@ -174,6 +213,19 @@ class actor{
 				}
 			}
 			fclose($handle);
+		}
+	}
+
+	static function isUsed($id){
+		global $bdd;
+		$req = $bdd->prepare('select count(*) from content_actor where idActor=:id');
+		$req->bindParam(":id", $id, PDO::PARAM_INT);
+		$result=$req->execute();
+		if ($result==true) {
+			$data = $req->fetch();
+			return $data[0];
+		}else{
+			return -1;
 		}
 	}
 }
